requirements.txt
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
contextlib2==21.6.0
cryptography==37.0.2
idna==3.3
oauthlib==3.2.0
pyaml==21.10.1
pycparser==2.21
PyJWT==2.4.0
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
schema==0.7.5
urllib3==1.26.9
